openai-web-app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.3
azure-identity==1.12.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docx2python==2.7.3
docx2txt==0.8
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
Flask==1.1.4
Flask-Sockets==0.2.1
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==2.11.3
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==0.8.4
msal==1.21.0
msal-extensions==1.0.0
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
openai==0.27.3
overrides==7.3.1
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.14.0
portalocker==2.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.3.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tiktoken==0.3.3
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==1.0.1
widgetsnbextension==4.0.7
yarl==1.8.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
